requirements.txt
discord.py
pytest
protobuf
